backend/requirements.txt
fastapi
uvicorn
yt-dlp
requests
openai
pymongo
python-dotenv
youtube_transcript_api
nltk
pytesseract 
pdf2image 
pillow
PyMuPDF
google-generativeai
python-jose
python-multipart
passlib[bcrypt]
pydantic
pydantic[email]
motor
